requirements.txt
openai
pydantic
python-dotenv
Jinja2
ruff
